app/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
starlette-graphene3==0.6.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
